+++ src/scope/repositories/source.js
@@ -1,10 +1,7 @@
 /** @flow */
 import * as path from 'path';
 import Repository from '../repository';
-import { SourceNotFound } from '../exceptions';
 import { BIT_SOURCES_DIRNAME } from '../../constants';
-import InvalidBit from '../../consumer/component/exceptions/invalid-bit';
-import Bit from '../../consumer/s';
 import { BitId } from '../../bit-id';
 import { listDirectories, rmDir, empty } from '../../utils';
 
@@ -13,47 +10,12 @@ export default class Source extends Repository {
     return path.join(super.getPath(), BIT_SOURCES_DIRNAME);
   }
 
-  getBitPath(bitName: string) {
-    return path.join(this.getPath(), bitName);
-  }
-
-  getPartial(name: string) {
-    // @TODO - partial bit does not exist anymore, implement something else to load bits in the sources dir
-    // return PartialBit.load(path.join(this.getPath(), name), name, this.scope.name);
-  }
-
-  setSource(bit: Bit, dependencies: Bit[]): Promise<Bit> {
-    if (!bit.validate()) throw new InvalidBit();
-    return bit
-      .cd(this.composeSourcePath(bit.getId()))
-      .write(true)
-      .then(() => this.scope.sourcesMap.setBit(bit.getId(), dependencies))
-      .then(() => bit);
-  }
-
   listVersions(bitId: BitId): number[] {
     return listDirectories(this.composeVersionsPath(bitId.name, bitId.box)).map(version => parseInt(version));
   }
 
   resolveVersion(id: BitId) {
     return id.getVersion().resolve(this.listVersions(id));
-  }
-
-  loadSource(id: BitId): Promise<Bit> {
-    try {
-      const version = this.resolveVersion(id);
-      return Bit.load(
-        this.composeSourcePath({
-          name: id.name,
-          box: id.box,
-          version
-        }),
-        id.name,
-        this.scope.name
-      );
-    } catch (err) {
-      throw new SourceNotFound(id);
-    }
   }
 
   isBoxEmpty(box: string) {