+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- move a re-imported component to a new location when `bit import --prefix` is used
 - fix commit and export issues when dealing with more than 500 components
 - support move command for moving files/directories of a component to a new location
 - fix export of large amount of data